# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Phising Simulator API",
        "contact": {},
        "version": "0.1"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/api/history": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "사용자의 과거 시뮬레이션(통화/채팅) 기록 목록을 최신순으로 반환합니다.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "API (Protected)"
                ],
                "summary": "사용자 통화 기록 조회",
                "responses": {
                    "200": {
                        "description": "history: [기록 배열]",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.HistoryResponse"
                        }
                    },
                    "401": {
                        "description": "인증 실패",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "DB 조회 실패 등 서버 오류",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/history/audio/{filename}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "특정 통화 기록의 오디오 파일(.mp3)을 재생합니다.\n\u003cbr\u003e **[인증]** Header에 `Authorization: Bearer ...`를 넣거나, URL 파라미터 `?token=...`을 사용하세요.",
                "produces": [
                    "audio/mpeg"
                ],
                "tags": [
                    "API (Protected)"
                ],
                "summary": "녹음된 오디오 파일 스트리밍",
                "parameters": [
                    {
                        "type": "string",
                        "description": "오디오 파일명 (예: session_uuid.mp3)",
                        "name": "filename",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "JWT 토큰 (Header 사용 불가 시)",
                        "name": "token",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "오디오 바이너리 데이터",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "401": {
                        "description": "인증 실패",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "해당 파일을 찾을 수 없음",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/profile": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "인증된 사용자의 프로필 정보를 조회합니다. (JWT 필요)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "API (Protected)"
                ],
                "summary": "프로필 조회 (Profile)",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "properties": {
                                " username": {
                                    "type": "string"
                                },
                                "message": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "인증 토큰 누락 또는 만료",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/login": {
            "post": {
                "description": "사용자명과 비밀번호로 로그인하고 JWT 토큰을 발급받습니다.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "로그인 (Login)",
                "parameters": [
                    {
                        "description": "로그인 요청 정보",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_handler.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.LoginSuccessResponse"
                        }
                    },
                    "400": {
                        "description": "잘못된 요청",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "인증 실패 (자격 증명 오류)",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "서버 내부 오류",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/signup": {
            "post": {
                "description": "새로운 사용자 계정을 생성합니다.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "회원가입 (Signup)",
                "parameters": [
                    {
                        "description": "회원가입 요청 정보",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_handler.SignupRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/ws/simulation": {
            "get": {
                "description": "지정된 시나리오와 모드로 실시간 시뮬레이션을 위한 WebSocket 연결을 시작합니다.\n\u003cbr\u003e\n**[중요]** 이것은 표준 HTTP API가 아닙니다. `ws://` 또는 `wss://` 스킴을 사용해야 합니다.\n**인증:** WebSocket 연결 시에는 HTTP Header를 사용할 수 없으므로, **Query Parameter(`token`)**로 JWT를 전달해야 합니다.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Simulation (WebSocket)"
                ],
                "summary": "보이스피싱 시뮬레이션 시작 (WebSocket)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer 토큰 (접두사 없이 토큰 값만 입력)",
                        "name": "token",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "시나리오 키 (예: loan_scam, institution_impersonation)",
                        "name": "scenario",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "모드 선택 (text: 텍스트 채팅, voice: 실시간 음성 통화)",
                        "name": "mode",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "101": {
                        "description": "Switching Protocols",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "잘못된 파라미터",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "인증 실패",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "PishingSimulator_SecurityProject_internal_models.Record": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "file_path": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "scenario": {
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "PishingSimulator_SecurityProject_internal_models.UserProfile": {
            "type": "object",
            "properties": {
                "age": {
                    "type": "integer"
                },
                "gender": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "internal_handler.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "에러 원인 및 설명"
                }
            }
        },
        "internal_handler.HistoryResponse": {
            "type": "object",
            "properties": {
                "history": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/PishingSimulator_SecurityProject_internal_models.Record"
                    }
                }
            }
        },
        "internal_handler.LoginRequest": {
            "type": "object",
            "properties": {
                "password": {
                    "type": "string",
                    "example": "password123"
                },
                "username": {
                    "type": "string",
                    "example": "my_user"
                }
            }
        },
        "internal_handler.LoginSuccessResponse": {
            "type": "object",
            "properties": {
                "token": {
                    "type": "string",
                    "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                }
            }
        },
        "internal_handler.SignupRequest": {
            "type": "object",
            "properties": {
                "password": {
                    "type": "string",
                    "example": "password123"
                },
                "profile": {
                    "$ref": "#/definitions/PishingSimulator_SecurityProject_internal_models.UserProfile"
                },
                "username": {
                    "type": "string",
                    "example": "new_user"
                }
            }
        },
        "internal_handler.SuccessResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "User created successfully"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Bearer 토큰 형식, Bearer {token}",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}